add confirmation to delete product

+++ client/src/components/homePageComponents/stock/StockSearch.jsx
@@ -24,15 +24,6 @@ const StockSearch = () => {
     const [searchQuery, setSearchQuery] = useState('');
     const [deleteId, setDeleteId] = useState('');
 
-    const [selectedCategory, setSelectedCategory] = useState('');
-    const [selectedType, setSelectedType] = useState('');
-    const [selectedCompany, setSelectedCompany] = useState('');
-    const [selectedStatus, setSelectedStatus] = useState('');
-    const [hasDiscount, setHasDiscount] = useState('');
-    const [minPrice, setMinPrice] = useState('');
-    const [maxPrice, setMaxPrice] = useState('');
-
-
     const [filteredProducts, setFilteredProducts] = useState([]);
     const [currentPage, setCurrentPage] = useState(1);
     const itemsPerPage = 400;
@@ -49,7 +40,6 @@ const StockSearch = () => {
 
     const allProducts = useSelector(state => state.saleItems.allProducts);
     const companyData = useSelector(state => state.companies.companyData);
-    const supplierData = useSelector(state => state.suppliers.supplierData);
     const categoryData = useSelector(state => state.categories.categoryData);
     const typeData = useSelector(state => state.types.typeData);
 
@@ -126,6 +116,10 @@ const StockSearch = () => {
         setIsButtonLoading(true);
         setDeleteId(id);
 
+        const confirm = window.confirm('Are You sure to want to delete this product? This Action cannot be undone!')
+
+        if(!confirm) return
+        
         try {
             const response = await config.deleteProduct(id);
             if (response) {
@@ -156,75 +150,17 @@ const StockSearch = () => {
     }, []);
 
     useEffect(() => {
-        let results = allProducts || [];
-
+        let results = allProducts;
         if (searchQuery) {
-            results = results.filter(
-                (p) =>
-                    p.productName.toLowerCase().includes(searchQuery.toLowerCase()) ||
-                    p.productCode?.toLowerCase().includes(searchQuery.toLowerCase())
+            results = allProducts?.filter(
+                (product) =>
+                    product.productName.toLowerCase().includes(searchQuery.toLowerCase()) ||
+                    product.productCode?.toLowerCase().includes(searchQuery.toLowerCase())
             );
         }
-
-        if (selectedCategory) {
-            results = results.filter(
-                (p) => p.categoryDetails[0]?.categoryName === selectedCategory
-            );
-        }
-
-        if (selectedType) {
-            results = results.filter(
-                (p) => p.typeDetails[0]?.typeName === selectedType
-            );
-        }
-
-        if (selectedCompany) {
-            results = results.filter(
-                (p) => p.vendorSupplierDetails[0]?.supplierName === selectedCompany
-            );
-        }
-
-        if (selectedStatus) {
-            if (selectedStatus === "Out of Stock") {
-                results = results.filter((p) => p.productTotalQuantity <= 0);
-            } else {
-                results = results.filter((p) => p.status === selectedStatus);
-            }
-        }
-
-        if (hasDiscount) {
-            if (hasDiscount === "with") {
-                results = results.filter((p) => Number(p.productDiscountPercentage) > 0);
-            } else {
-                results = results.filter((p) => !p.productDiscountPercentage || Number(p.productDiscountPercentage) === 0);
-            }
-        }
-
-        if (minPrice || maxPrice) {
-            results = results.filter((p) => {
-                const price = Number(p.salePriceDetails?.[0]?.salePrice1 || 0);
-                return (
-                    (!minPrice || price >= Number(minPrice)) &&
-                    (!maxPrice || price <= Number(maxPrice))
-                );
-            });
-        }
-        console.log('results', results)
-
         setFilteredProducts(results);
         setCurrentPage(1);
-    }, [
-        searchQuery,
-        selectedCategory,
-        selectedType,
-        selectedCompany,
-        selectedStatus,
-        hasDiscount,
-        minPrice,
-        maxPrice,
-        allProducts
-    ]);
-
+    }, [searchQuery, allProducts]);
 
     const indexOfLastItem = currentPage * itemsPerPage;
     const indexOfFirstItem = indexOfLastItem - itemsPerPage;
@@ -235,7 +171,7 @@ const StockSearch = () => {
             "Code", "Name", "Type", "Pack", "Company", "Vendor", "product Discount Percentage", "Category", "product Purchase Price", "Sale Price1", "Sale Price2", "Sale Price3", "Sale Price4", "Total Qty", "status"
         ];
 
-        const rows = allProducts?.map(product => [
+        const rows = currentProducts.map(product => [
             product.productCode,
             product.productName,
             product.typeDetails[0]?.typeName,
@@ -268,7 +204,7 @@ const StockSearch = () => {
 
     return !isEdit ? (
         <div className='bg-white rounded-lg'>
-            <div className="w-full px-5 py-">
+            <div className="w-full px-5 py-5">
                 <h2 className="text-lg text-center pt-2 font-semibold mb-2">Search for Stock</h2>
                 <div className="text-xs text-red-500 mb-2 text-center">
                     {error && <p>{error}</p>}
@@ -285,77 +221,6 @@ const StockSearch = () => {
                         onChange={(e) => setSearchQuery(e.target.value)}
                         ref={inputRef}
                     />
-                    <div className="flex flex-wrap gap-3 my-3 items-center text-xs">
-                        <select
-                            value={selectedCategory}
-                            onChange={(e) => setSelectedCategory(e.target.value)}
-                            className="border p-1 rounded"
-                        >
-                            <option value="">All Categories</option>
-                            {categoryData?.map((cat) => (
-                                <option key={cat._id} value={cat.categoryName}>{cat.categoryName}</option>
-                            ))}
-                        </select>
-
-                        <select
-                            value={selectedType}
-                            onChange={(e) => setSelectedType(e.target.value)}
-                            className="border p-1 rounded"
-                        >
-                            <option value="">All Types</option>
-                            {typeData?.map((type) => (
-                                <option key={type._id} value={type.typeName}>{type.typeName}</option>
-                            ))}
-                        </select>
-
-                        <select
-                            value={selectedCompany}
-                            onChange={(e) => setSelectedCompany(e.target.value)}
-                            className="border p-1 rounded"
-                        >
-                            <option value="">All Suppliers</option>
-                            {supplierData?.map((com) => (
-                                <option key={com._id} value={com.supplierName}>{com.supplierName}</option>
-                            ))}
-                        </select>
-
-
-                        <div className="flex items-center gap-1">
-                            <label>Price:</label>
-                            <input
-                                type="number"
-                                placeholder="Min"
-                                value={minPrice}
-                                onChange={(e) => setMinPrice(e.target.value)}
-                                className="w-16 border p-1 rounded"
-                            />
-                            <span>-</span>
-                            <input
-                                type="number"
-                                placeholder="Max"
-                                value={maxPrice}
-                                onChange={(e) => setMaxPrice(e.target.value)}
-                                className="w-16 border p-1 rounded"
-                            />
-                        </div>
-
-                        <button
-                            onClick={() => {
-                                setSelectedCategory('');
-                                setSelectedType('');
-                                setSelectedCompany('');
-                                setSelectedStatus('');
-                                setHasDiscount('');
-                                setMinPrice('');
-                                setMaxPrice('');
-                                setSearchQuery('');
-                            }}
-                            className="text-xs text-white bg-red-500 hover:bg-red-400 px-2 py-1 rounded"
-                        >
-                            Clear Filters
-                        </button>
-                    </div>
-
                     <button
                         onClick={exportToCSV}
                         className="bg-green-500 hover:bg-green-600 text-white py-1 px-3 rounded text-xs"
@@ -390,7 +255,7 @@ const StockSearch = () => {
                                         <td className="px-1 py-1">{product.productPack}</td>
                                         <td className="px-1 py-1">{product.companyDetails[0]?.companyName}</td>
                                         <td className="px-1 py-1">{product.vendorSupplierDetails[0]?.supplierName || product.vendorCompanyDetails[0]?.companyName}</td>
-                                        <td className="px-1 py-1">{product.categoryDetails[0]?.categoryName}</td>
+                                        <td className="px-1 py-1">{product.categoryDetails[0]?.productName}</td>
                                         <td className="px-1 py-1">{product.salePriceDetails[0]?.salePrice1}</td>
                                         <td className="px-1 py-1">{Math.ceil(product.productTotalQuantity / product.productPack)}</td>
                                         <td className="px-1 py-1">{Math.ceil(product.productTotalQuantity)}</td>
